+++ src/analysis/fluxes.jl
@@ -25,10 +25,8 @@ function make_Ie_top_file(directory_to_process; max_bytes::Real = 512 * 1024^2)
     E_centers = coord.E_centers
     ΔE        = coord.ΔE
     μ_lims    = coord.μ_lims
-
-    ## Beam weights
-    θ_lims = acosd.(μ_lims)
-    Ω_beam = beam_weight(θ_lims)   # [n_μ]
+    μ_center  = coord.μ_center
+    Ω_beam    = coord.beam_weights
 
     n_z, n_μ, n_t, n_E = coord.n_z, coord.n_μ, coord.n_t, coord.n_E
 
@@ -50,7 +48,7 @@ function make_Ie_top_file(directory_to_process; max_bytes::Real = 512 * 1024^2)
 
         pa_v = defVar(ds, "pitch_angle", Float64, ("pitch_angle",);
                       attrib=["units" => "1", "long_name" => "cosine of pitch angle"])
-        pa_v[:] = mu_avg(θ_lims)
+        pa_v[:] = μ_center
 
         t_v = defVar(ds, "time", Float64, ("time",);
                      attrib=["units" => "s", "long_name" => "simulation time"])
@@ -129,13 +127,7 @@ function make_current_file(directory_to_process; max_bytes::Real = 512 * 1024^2)
     t         = coord.t
     z         = coord.h_atm
     E_centers = coord.E_centers
-    μ_lims    = coord.μ_lims
-
-    θ_lims   = acosd.(μ_lims)
-    μ_center = mu_avg(θ_lims)
-
-    ## Elementary charge
-    q_e = 1.602176620898e-19  # C
+    μ_center  = coord.μ_center
 
     n_z, n_μ, n_t = coord.n_z, coord.n_μ, coord.n_t
     J_up      = zeros(n_z, n_t)
@@ -149,10 +141,10 @@ function make_current_file(directory_to_process; max_bytes::Real = 512 * 1024^2)
             flux_block   = dropdims(sum(Ie_chunk[:, i_μ, :, :], dims=3), dims=3)   # [n_z, n_t_chunk]
             energy_block = dropdims(sum(Ie_chunk[:, i_μ, :, :] .* reshape(E_centers, 1, 1, :), dims=3), dims=3)
             if μ_center[i_μ] > 0
-                J_up[:, t_range]   .+= q_e * abs(μ_center[i_μ]) .* flux_block
+                J_up[:, t_range]   .+= qₑ * abs(μ_center[i_μ]) .* flux_block
                 IeE_up[:, t_range] .+= abs(μ_center[i_μ]) .* energy_block
             else
-                J_down[:, t_range]   .+= q_e * abs(μ_center[i_μ]) .* flux_block
+                J_down[:, t_range]   .+= qₑ * abs(μ_center[i_μ]) .* flux_block
                 IeE_down[:, t_range] .+= abs(μ_center[i_μ]) .* energy_block
             end
         end