+++ src/constants.jl
@@ -1,0 +1,5 @@
+# Physical constants
+const qₑ = 1.602176634e-19    # elementary charge [C]
+const mₑ = 9.1093837015e-31   # electron mass [kg]
+const kB = 1.380649e-23       # Boltzmann constant [J/K]
+const c₀ = 299792458.0        # speed of light in vacuum [m/s]